# docs/swagger.yaml
basePath: /v1
definitions:
...
  models.Doorlock:
    properties:
      activeState:
        type: string
      blockId:
        type: string
      connectState:
        type: string
      description:
        type: string
      doorSerialId:
        type: string
      doorState:
        type: string
      doorlockAddress:
        type: string
      floorId:
        type: string
      gatewayId:
        type: string
      id:
        type: integer
      lastOpenTime:
        type: integer
      location:
        type: string
      lockState:
        type: string
      roomId:
        type: string
      schedulers:
        items:
          $ref: '#/definitions/models.Scheduler'
        type: array
    type: object
  models.DoorlockCmd:
    properties:
      duration:
        type: string
      id:
        type: string
      state:
        type: string
    type: object
  models.DoorlockStatus:
    properties:
      connectState:
        type: string
      doorState:
        type: string
      doorlockAddress:
        type: string
      gatewayId:
        type: string
      id:
        type: string
      lockState:
        type: string
    type: object
...
  models.Scheduler:
    properties:
      amount:
        type: integer
      base:
        type: string
      capacity:
        type: integer
      classId:
        type: string
      className:
        type: string
      customerId:
        type: string
      doorId:
        description: DoorID          *string `gorm:"type:varchar(256);" json:"doorId"`
        type: integer
      employeeId:
        type: string
      endClassTime:
        type: integer
      endDate:
        type: string
      id:
        type: integer
      lecturerId:
        type: string
      lecturerName:
        type: string
      role:
        description: value in ["employee", "student", "customer"]
        type: string
      roomId:
        type: string
      roomName:
        type: string
      roomRow:
        type: string
      startClassTime:
        type: integer
      startDate:
        type: string
      status:
        type: string
      studentId:
        type: string
      userId:
        type: string
      weekDay:
        type: integer
    type: object
...
  models.SwagCreateDoorlock:
    properties:
      areaId:
        type: integer
      description:
        type: string
      doorlockAddress:
        type: string
      gatewayId:
        type: integer
      location:
        type: string
      schedulerId:
        type: integer
    type: object
...
  models.SwagUpdateDoorlock:
    properties:
      activeState:
        type: string
      blockId:
        type: string
      description:
        type: string
      doorSerialId:
        type: string
      doorlockAddress:
        type: string
      floorId:
        type: string
      gatewayId:
        type: string
      id:
        type: integer
      location:
        type: string
      roomId:
        type: string
    type: object
...
  models.SwaggerDoorlockCmd:
    properties:
      duration:
        type: string
      id:
        type: string
      state:
        type: string
    type: object
...
  utils.ErrorResponse:
    properties:
      errorMsg:
        type: string
      msg:
        type: string
      statusCode:
        type: integer
    type: object
host: http://iot.hcmue.space:8002
info:
  contact:
    name: Mr. Khai
  description: This is API document for DMS backend server
  title: Backend API
  version: "1.0"
paths:
...
  /v1/doorlock:
    delete:
      consumes:
      - application/json
      description: Delete doorlock using "id" field. Send deleted info to MQTT broker
      parameters:
      - description: Doorlock Delete payload
        in: body
        name: data
        required: true
        schema:
          allOf:
          - type: object
          - properties:
              id:
                type: integer
            type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Delete Doorlock By ID
    patch:
      consumes:
      - application/json
      description: Update doorlock, must have "gatewayId" and "doorlockAddress" field.
        Send updated info to MQTT broker
      parameters:
      - description: Fields need to update a doorlock
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.SwagUpdateDoorlock'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Update Doorlock By Doorlock Address and GatewayID
    post:
      consumes:
      - application/json
      description: Create doorlock. Send created info to MQTT broker
      parameters:
      - description: Fields need to create a doorlock
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.SwagCreateDoorlock'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Doorlock'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Create Doorlock
  /v1/doorlock/{id}:
    get:
      description: find doorlock info by doorlock id
      parameters:
      - description: Doorlock ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Doorlock'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Find Doorlock By ID
  /v1/doorlock/cmd:
    patch:
      consumes:
      - application/json
      description: Update doorlock state, must have "id" field. Send updated info
        to MQTT broker
      parameters:
      - description: Fields need to update a doorlock state
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.SwaggerDoorlockCmd'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Update Doorlock State By ID
  /v1/doorlock/state/cmd:
    patch:
      consumes:
      - application/json
      description: Send command lock/unlock forever to Doorlock by serialID and save
        its lock state
      parameters:
      - description: Fields need to update a doorlock state
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.DoorlockCmd'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Send command lock/unlock forever to Doorlock and update doorlock's
        lock state
  /v1/doorlock/status/{id}:
    get:
      description: Get doorlock status by doorlock id
      parameters:
      - description: Doorlock ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DoorlockStatus'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Get Doorlock Status By ID
...
  /v1/doorlocks:
    get:
      description: find all doorlocks info
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/models.Doorlock'
              type: array
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Find All Doorlock
...
swagger: "2.0"